requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
schedule==1.2.0
pandas==2.0.3
openpyxl==3.1.2
plotly==5.17.0
requests==2.31.0
python-dateutil==2.8.2
